fix(suggestion-mapper): correct type casting in articles size mapping
- Corrected the type casting in `SuggestionMapper` to ensure the size of articles is cast to `Long`.
- Added `@EntityGraph` annotation to `findTop10ByOrderByCreatedAtDesc` method in `SuggestionRepository` to optimize fetching related entities.

These changes fix a type casting issue in the `SuggestionMapper` and improve the performance of fetching suggestions by including related entities in the query.

+++ server/src/main/java/com/barataribeiro/sentinelofliberty/builders/SuggestionMapper.java
@@ -24,7 +24,7 @@ public class SuggestionMapper {
             protected void configure() {
                 using(ctx -> {
                     Set<?> articles = (Set<?>) ctx.getSource();
-                    return articles == null ? 0L : articles.size();
+                    return articles == null ? 0L : (long) articles.size();
                 }).map(source.getArticles(), destination.getArticlesWritten());
             }
         });
